Fix duplicated counts when reconnecting
The UI does not show the multiple connected service buses in the
collection, but does show ever-increasing topic/queue counts when
re-connecting without quitting the app.

This patch resets the state when a new connection is made to avoid
showing the old connection queue/topic counts added to the new
connection queue/topic counts.